+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Features
 
+* [#684](https://github.com/ruby-grape/grape-swagger/pull/684): Add response headers - [@jdmurphy](https://github.com/jdmurphy).
 * Your contribution here.
 
 #### Fixes